Move `model` to `data` module
